requirements.txt
absl-py==2.0.0
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
clean-text==0.6.0
click==8.1.7
clip-interrogator==0.6.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
datasets==2.15.0
deepspeed==0.12.4+02288bc1
diffusers==0.24.0
dill==0.3.7
distro==1.8.0
docker-pycreds==0.4.0
emoji==1.7.0
evaluate==0.4.1
exceptiongroup==1.2.0
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.9.0
fire==0.5.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
gradio==4.7.1
gradio_client==0.7.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
inflate64==1.0.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
multivolumefile==0.2.3
natsort==8.4.0
networkx==3.0
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.2
open-clip-torch==2.23.0
openai==1.3.7
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.1.3
Pillow==9.3.0
protobuf==4.25.1
psutil==5.9.6
py-cpuinfo==9.0.0
py7zr==0.20.8
pyarrow==14.0.1
pyarrow-hotfix==0.6
pybcj==1.0.2
pycodestyle==2.11.1
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
pyppmd==1.1.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzstd==0.15.9
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
responses==0.18.0
rich==13.7.0
rouge-score==0.1.2
rpds-py==0.13.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.38.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
sklearn==0.0.post11
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
termcolor==2.3.0
textsum==0.2.0
texttable==1.7.0
threadpoolctl==3.2.0
timm==0.9.12
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch>=2.1.1
torchaudio>=2.1.1
torchvision>=0.16.1
tqdm==4.66.1
transformers @ git+https://github.com/huggingface/transformers@510270af34994c24f2f8792d2afd74114c70d736
triton==2.1.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
wandb==0.16.0
wcwidth==0.2.12
websockets==11.0.3
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
